move the default Defaults implementation to a separate class that is only loaded when needed, this makes it possible to exclude the json-smart dependency (see https://code.google.com/p/json-path/issues/detail?id=68)
